move as_awaitable into its own header
